# implementation/openapi.json
{
  "components": {
    "schemas": {
...
      "AdminAccountListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AdminAccountSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "organisationId": {
            "title": "Organisationid",
            "type": "string"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "pageSize": {
            "title": "Pagesize",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "page",
          "pageSize",
          "total",
          "organisationId"
        ],
        "title": "AdminAccountListResponse",
        "type": "object"
      },
...
      "AdminFacilityListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AdminFacilityItem"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "AdminFacilityListResponse",
        "type": "object"
      },
...
      "AdminHomeSummaryResponse": {
        "properties": {
          "backendAvailable": {
            "title": "Backendavailable",
            "type": "boolean"
          },
          "inactiveWorkerCount": {
            "title": "Inactiveworkercount",
            "type": "integer"
          },
          "organisationId": {
            "title": "Organisationid",
            "type": "string"
          },
          "pendingFirstLoginCount": {
            "title": "Pendingfirstlogincount",
            "type": "integer"
          },
          "workerCount": {
            "title": "Workercount",
            "type": "integer"
          }
        },
        "required": [
          "organisationId",
          "workerCount",
          "pendingFirstLoginCount",
          "inactiveWorkerCount",
          "backendAvailable"
        ],
        "title": "AdminHomeSummaryResponse",
        "type": "object"
      },
...
      "DevelopmentTokenRequest": {
        "properties": {
          "account_id": {
            "anyOf": [
              {
                "maxLength": 64,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Id"
          },
          "email": {
            "anyOf": [
              {
                "maxLength": 320,
                "minLength": 3,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "identifier": {
            "anyOf": [
              {
                "maxLength": 320,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Identifier"
          },
          "password": {
            "minLength": 1,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "password"
        ],
        "title": "DevelopmentTokenRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RegisterWorkerRequest": {
        "properties": {
          "communityRequestsEnabled": {
            "title": "Communityrequestsenabled",
            "type": "boolean"
          },
          "displayName": {
            "maxLength": 255,
            "minLength": 2,
            "title": "Displayname",
            "type": "string"
          },
          "email": {
            "maxLength": 320,
            "minLength": 3,
            "title": "Email",
            "type": "string"
          },
          "emergencyRequestsEnabled": {
            "title": "Emergencyrequestsenabled",
            "type": "boolean"
          },
          "facilityId": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Facilityid",
            "type": "string"
          },
          "idempotencyKey": {
            "maxLength": 128,
            "minLength": 8,
            "title": "Idempotencykey",
            "type": "string"
          },
          "organisationId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Organisationid"
          },
          "otherProfessionDescription": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Otherprofessiondescription"
          },
          "profession": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Profession",
            "type": "string"
          },
          "role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role"
          },
          "temporaryPassword": {
            "maxLength": 128,
            "minLength": 12,
            "title": "Temporarypassword",
            "type": "string"
          }
        },
        "required": [
          "displayName",
          "email",
          "facilityId",
          "temporaryPassword",
          "idempotencyKey",
          "profession",
          "communityRequestsEnabled",
          "emergencyRequestsEnabled"
        ],
        "title": "RegisterWorkerRequest",
        "type": "object"
      },
      "RegisterWorkerResponse": {
        "properties": {
          "accountId": {
            "title": "Accountid",
            "type": "string"
          },
          "accountStatus": {
            "title": "Accountstatus",
            "type": "string"
          },
          "accountVersion": {
            "title": "Accountversion",
            "type": "integer"
          },
          "displayName": {
            "title": "Displayname",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "facilityId": {
            "title": "Facilityid",
            "type": "string"
          },
          "firstLoginRequired": {
            "title": "Firstloginrequired",
            "type": "boolean"
          },
          "identityProvider": {
            "title": "Identityprovider",
            "type": "string"
          },
          "professionalProfile": {
            "$ref": "#/components/schemas/ProfessionalProfileResponse"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "title": "Roles",
            "type": "array"
          }
        },
        "required": [
          "accountId",
          "displayName",
          "email",
          "roles",
          "facilityId",
          "accountStatus",
          "firstLoginRequired",
          "accountVersion",
          "identityProvider",
          "professionalProfile"
        ],
        "title": "RegisterWorkerResponse",
        "type": "object"
      },
...
      "SessionAuthorisationResponse": {
        "properties": {
          "accountId": {
            "title": "Accountid",
            "type": "string"
          },
          "accountStatus": {
            "title": "Accountstatus",
            "type": "string"
          },
          "accountVersion": {
            "title": "Accountversion",
            "type": "integer"
          },
          "displayName": {
            "title": "Displayname",
            "type": "string"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "firstLoginRequired": {
            "title": "Firstloginrequired",
            "type": "boolean"
          },
          "identityProvider": {
            "title": "Identityprovider",
            "type": "string"
          },
          "organisationId": {
            "title": "Organisationid",
            "type": "string"
          },
          "permittedWorkspaces": {
            "items": {
              "type": "string"
            },
            "title": "Permittedworkspaces",
            "type": "array"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "title": "Roles",
            "type": "array"
          },
          "workerFacilityId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workerfacilityid"
          }
        },
        "required": [
          "accountId",
          "displayName",
          "roles",
          "permittedWorkspaces",
          "accountStatus",
          "organisationId",
          "firstLoginRequired",
          "accountVersion",
          "identityProvider"
        ],
        "title": "SessionAuthorisationResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "NorthCare AI sync, administration, Reach community-request API, USSD simulator, and Africa's Talking USSD sandbox adapter",
    "title": "NorthCare AI Sync API",
    "version": "0.18.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/admin/accounts": {
      "get": {
        "operationId": "list_accounts_v1_admin_accounts_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 50,
              "minimum": 1,
              "title": "Pagesize",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "in": "query",
            "name": "facilityId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Facilityid"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminAccountListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Accounts",
        "tags": [
          "administration"
        ]
      },
      "post": {
        "operationId": "register_worker_v1_admin_accounts_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterWorkerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterWorkerResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register Worker",
        "tags": [
          "administration"
        ]
      }
    },
...
    "/v1/admin/facilities": {
      "get": {
        "operationId": "list_facilities_v1_admin_facilities_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminFacilityListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Facilities",
        "tags": [
          "administration"
        ]
      }
    },
    "/v1/admin/home": {
      "get": {
        "operationId": "admin_home_v1_admin_home_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminHomeSummaryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Home",
        "tags": [
          "administration"
        ]
      }
    },
...
    "/v1/auth/session": {
      "get": {
        "operationId": "get_session_authorisation_v1_auth_session_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionAuthorisationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Session Authorisation",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/development/auth/token": {
      "post": {
        "operationId": "issue_development_token_v1_development_auth_token_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DevelopmentTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DevelopmentTokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Issue Development Token",
        "tags": [
          "development-auth"
        ]
      }
    },
    "/v1/devices/register": {
      "post": {
        "operationId": "register_v1_devices_register_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceRegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceRegisterResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register",
        "tags": [
          "devices"
        ]
      }
    },
...
    "/v1/sync/changes": {
      "get": {
        "operationId": "sync_changes_v1_sync_changes_get",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncPullResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sync Changes",
        "tags": [
          "sync"
        ]
      }
    },
    "/v1/sync/conflicts": {
      "get": {
        "operationId": "conflicts_v1_sync_conflicts_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Conflicts",
        "tags": [
          "sync"
        ]
      }
    },
    "/v1/sync/conflicts/{conflict_id}/resolve": {
      "post": {
        "operationId": "resolve_v1_sync_conflicts__conflict_id__resolve_post",
        "parameters": [
          {
            "in": "path",
            "name": "conflict_id",
            "required": true,
            "schema": {
              "title": "Conflict Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConflictResolveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictResolveResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolve",
        "tags": [
          "sync"
        ]
      }
    },
    "/v1/sync/push": {
      "post": {
        "operationId": "sync_push_v1_sync_push_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncPushRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncPushResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sync Push",
        "tags": [
          "sync"
        ]
      }
    },
...
  }
}